requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.4
django-extensions==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==1.2.1
ecdsa==0.18.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
Pillow==9.3.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.13
wrapt==1.14.1
